refactor: split wasm.rs
