+++ crates/bitwarden-shared-unlock/src/wasm/mod.rs
@@ -1,0 +1,8 @@
+mod drivers;
+mod follower;
+mod leader;
+mod sender;
+
+pub use drivers::WasmDriverModule;
+pub use follower::SharedUnlockFollower;
+pub use leader::SharedUnlockLeader;
